requirements.txt
aiodns==3.0.0
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.1.0
ccxt==2.2.40
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
dhooks==1.1.4
fastapi==0.88.0
fire==0.4.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.2
httptools==0.5.0
httpx==0.23.1
idna==3.4
loguru==0.6.0
multidict==6.0.2
pycares==4.2.2
pycparser==2.21
pydantic==1.10.2
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
termcolor==2.1.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
yarl==1.8.1
